# docs/openapi.yaml
openapi: 3.1.0
info:
  title: FreshMart Agentic Analytics API
  description: "\n# FreshMart Agentic Analytics API\n\nWelcome to the **FreshMart\
    \ Agentic Analytics Platform** enterprise API documentation.\n\nFreshMart Agentic\
    \ Analytics provides secure, AI-powered natural-language analytics for FreshMart\
    \ enterprise business data across retail operations, customer relationship management,\
    \ human resources, inventory warehousing, purchasing, and financial performance.\n\
    \n---\n\n### Core Platform Capabilities\n\n* **AI-Powered Natural-Language Analytics:**\
    \ Inquire in plain English (e.g., *\"Why did profit decrease in August?\"* or\
    \ *\"Show monthly sales trend for the last 12 months\"*).\n* **LangGraph Multi-Agent\
    \ Orchestration:** Specialized agents collaborate autonomously:\n  * **Intent\
    \ Classifier:** Detects analytical goal, business domain, and inquiry category.\n\
    \  * **Permission Checker:** Enforces database table and column-level RBAC prior\
    \ to query generation.\n  * **Schema Planner:** Dynamically inspects schemas using\
    \ the Model Context Protocol (MCP).\n  * **SQL Generator & Corrector:** Synthesizes\
    \ read-only SQL via local `qwen2.5-coder:7b`.\n  * **SQL Validator:** Rigorous\
    \ AST parser preventing non-SELECT, destructive, or unauthorized queries.\n  *\
    \ **SQL Executor:** Safely runs validated SQL through MCP with row bounding and\
...
    \ `admin` / `AdminPassword123!`\n"
  contact:
...
paths:
...
  /api/v1/admin/users:
    get:
      tags:
      - User Management
      summary: List All Registered Application Users
      description: 'Retrieves a complete list of all FreshMart application user accounts,
        including their assigned role, granted permission codes, workforce employee
        linkage, and active status.


        **Authorization:** Requires `MANAGE_USERS` permission (granted to `ADMIN`).'
      operationId: listAllApplicationUsers
      responses:
        '200':
          description: List of all user accounts successfully retrieved.
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/UserResponse'
                type: array
                title: Response 200 Listallapplicationusers
        '401':
          description: 'Unauthorized: Missing, invalid, or expired JWT access token.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '403':
          description: 'Forbidden: User lacks the required ''MANAGE_USERS'' permission.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      security:
      - bearerAuth: []
      - OAuth2PasswordBearer: []
    post:
      tags:
      - User Management
      summary: Provision New Application User Account
      description: 'Provisions a new FreshMart user account with secure bcrypt password
        hashing and associates it with an existing application role.


        **Authorization:** Requires `MANAGE_USERS` permission (granted to `ADMIN`).'
      operationId: createApplicationUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
        required: true
      responses:
        '201':
          description: User account created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: 'Bad Request: Username or email already exists in system.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '401':
          description: 'Unauthorized: Missing, invalid, or expired JWT access token.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '403':
          description: 'Forbidden: User lacks the required ''MANAGE_USERS'' permission.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '404':
          description: 'Not Found: The specified role_id does not exist.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '422':
          description: 'Validation Error: Invalid payload attributes.'
      security:
      - bearerAuth: []
      - OAuth2PasswordBearer: []
  /api/v1/admin/users/{user_id}/status:
    patch:
      tags:
      - User Management
      summary: Activate or Deactivate User Account
      description: 'Enables or disables an application user account. Deactivated accounts
        are immediately blocked from logging in and their active JWT sessions are
        rejected at the authentication barrier.


        **Authorization:** Requires `MANAGE_USERS` permission (granted to `ADMIN`).'
      operationId: updateUserAccountStatus
      security:
      - bearerAuth: []
      - OAuth2PasswordBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
          title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserStatusUpdate'
      responses:
        '200':
          description: User account status updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: 'Unauthorized: Missing, invalid, or expired JWT access token.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '403':
          description: 'Forbidden: User lacks the required ''MANAGE_USERS'' permission.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '404':
          description: 'Not Found: User with specified user_id not found.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '422':
          description: 'Validation Error: Invalid user_id or payload.'
  /api/v1/admin/users/{user_id}/role:
    patch:
      tags:
      - User Management
      summary: Change User Application Role & Privileges
      description: 'Reassigns a user''s application role, dynamically changing their
        granted RBAC permission codes and data domain access across the platform.


        **Authorization:** Requires `MANAGE_ROLES` permission (granted to `ADMIN`).'
      operationId: updateUserApplicationRole
      security:
      - bearerAuth: []
      - OAuth2PasswordBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
          title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRoleUpdate'
      responses:
        '200':
          description: User role updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: 'Unauthorized: Missing, invalid, or expired JWT access token.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '403':
          description: 'Forbidden: User lacks the required ''MANAGE_ROLES'' permission.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '404':
          description: 'Not Found: User or Role ID not found.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '422':
          description: 'Validation Error: Invalid user_id or role_id.'
  /api/v1/admin/roles:
    get:
      tags:
      - Role Management
      summary: List Application Roles & Granted Permission Codes
      description: 'Returns all 7 FreshMart application roles (ADMIN, CEO, HR_MANAGER,
        ERP_MANAGER, SALES_MANAGER, INVENTORY_MANAGER, FINANCE_MANAGER) along with
        their granular permission codes.


        **Authorization:** Requires `MANAGE_ROLES` permission (granted to `ADMIN`).'
      operationId: listApplicationRolesAndPermissions
      responses:
        '200':
          description: Roles and assigned permission codes successfully retrieved.
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/RoleResponse'
                type: array
                title: Response 200 Listapplicationrolesandpermissions
        '401':
          description: 'Unauthorized: Missing, invalid, or expired JWT access token.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '403':
          description: 'Forbidden: User lacks the required ''MANAGE_ROLES'' permission.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      security:
      - bearerAuth: []
      - OAuth2PasswordBearer: []
...
components:
  schemas:
...
    HTTPError:
      properties:
        detail:
          type: string
          title: Detail
          description: Human-readable explanation of why the request failed.
          example: 'Access Denied: You do not have permission ''VIEW_EMPLOYEE_SALARY''.'
      type: object
      required:
      - detail
      title: HTTPError
      description: Standard RFC 7807-compatible error response detail.
...
    RoleResponse:
      properties:
        role_id:
          type: integer
          title: Role Id
          description: Unique role identifier
          example: 5
        role_name:
          type: string
          title: Role Name
          description: Unique role name
          example: SALES_MANAGER
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
          description: Role purpose description
          example: Sales Manager. Customers, CRM, orders, and sales trends.
        permissions:
          items:
            type: string
          type: array
          title: Permissions
          description: List of permission codes assigned to this role
          example:
          - VIEW_CRM
          - VIEW_SALES
          - VIEW_INVENTORY
      type: object
      required:
      - role_id
      - role_name
      - permissions
      title: RoleResponse
      description: Application role representation with assigned permissions.
...
    UserCreate:
      properties:
        username:
          type: string
          maxLength: 60
          minLength: 3
          title: Username
          description: Unique username for the new account
          example: anita.finance
        email:
          type: string
          title: Email
          description: Unique email address
          example: anita.f@freshmart.local
        password:
          type: string
          minLength: 6
          title: Password
          description: Initial plain text password (hashed securely with bcrypt)
          example: SecurePass2026!
        full_name:
          type: string
          maxLength: 120
          title: Full Name
          description: Full display name of the user
          example: Anita Sharma
        role_id:
          type: integer
          title: Role Id
          description: 'Role ID to assign to this user (1: ADMIN, 2: CEO, 3: HR_MANAGER,
            4: ERP_MANAGER, 5: SALES_MANAGER, 6: INVENTORY_MANAGER, 7: FINANCE_MANAGER)'
          example: 7
        employee_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Employee Id
          description: Optional MySQL employee ID to associate
          example: E010
      type: object
      required:
      - username
      - email
      - password
      - full_name
      - role_id
      title: UserCreate
      description: Payload to provision a new application user account.
...
    UserResponse:
      properties:
        user_id:
          type: integer
          title: User Id
          description: Unique user identifier
          example: 1
        username:
          type: string
          title: Username
          description: Unique username
          example: admin
        email:
          type: string
          title: Email
          description: Email address
          example: admin@freshmart.local
        full_name:
          type: string
          title: Full Name
          description: Display name
          example: Technical Administrator
        role_id:
          type: integer
          title: Role Id
          description: Foreign key reference to assigned role
          example: 1
        role_name:
          type: string
          title: Role Name
          description: Name of assigned role
          example: ADMIN
        employee_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Employee Id
          description: Linked MySQL workforce ID
        is_active:
          type: boolean
          title: Is Active
          description: Whether user account is active
          example: true
        permissions:
          items:
            type: string
          type: array
          title: Permissions
          description: List of granular permission codes granted via role
          example:
          - MANAGE_USERS
          - MANAGE_ROLES
        created_at:
          type: string
          format: date-time
          title: Created At
          description: Account creation UTC timestamp
      type: object
      required:
      - user_id
      - username
      - email
      - full_name
      - role_id
      - role_name
      - is_active
      - permissions
      - created_at
      title: UserResponse
      description: Detailed user account representation.
    UserRoleUpdate:
      properties:
        role_id:
          type: integer
          title: Role Id
          description: ID of the new role to assign
          example: 5
      type: object
      required:
      - role_id
      title: UserRoleUpdate
      description: Payload to reassign a user's application role.
    UserStatusUpdate:
      properties:
        is_active:
          type: boolean
          title: Is Active
          description: Set true to enable account, false to deactivate account
          example: false
      type: object
      required:
      - is_active
      title: UserStatusUpdate
      description: Payload to activate or deactivate a user account.
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Authorization header using the Bearer scheme. Enter token directly
        or as 'Bearer <token>'.
    OAuth2PasswordBearer:
      type: oauth2
      description: 'OAuth2 Password Flow: Authenticate using username and password'
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login
...